separate lexer and symbols from symbol
