+++ README.md
@@ -5,7 +5,7 @@ example code
 ```cpp
 // main.cpp
 
-#include <tinylexer/symbol.hpp>
+#include <tinylexer/lexer.hpp>
 
 int main()
 {
